requirements-test.txt
py==1.4.34
pytest==3.2.2
pytest-django==3.1.2
